Add a verbose option to the bb-components script
When the `verbose` option is true, the script prints out the list of
changed Clojure namespaces  that was picked up, as well as the list of
namespaces due to which a certain component was picked up.

This is helpful to identify why a certain component was affected by a
seemingly unrelated change.