Add a stdin option to the bb-components script
This option allows the user to control the specific files used to
calculate the impact of the change. This is particularly useful to
remove files which are heavily used, if the user knows that the change
in those files is non-breaking and can be ignored.

Updated the README with examples, to explain the usage better.